requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple
--trusted-host https

amqp==5.2.0
    # via kombu
async-timeout==4.0.3
    # via redis
attrs==23.2.0
    # via zeep
billiard==4.2.0
    # via celery
celery[redis]==5.3.6
    # via -r requirements.in
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dnspython==2.6.1
    # via eventlet
eventlet==0.35.2
    # via -r requirements.in
fonttools==4.50.0
    # via matplotlib
greenlet==3.0.3
    # via eventlet
idna==3.6
    # via requests
importlib-resources==6.3.0
    # via matplotlib
isodate==0.6.1
    # via zeep
joblib==1.3.2
    # via
    #   pyod
    #   scikit-learn
kiwisolver==1.4.5
    # via matplotlib
kombu==5.3.5
    # via celery
llvmlite==0.42.0
    # via numba
lxml==5.2.1
    # via zeep
matplotlib==3.8.3
    # via
    #   pyod
    #   pythresh
    #   sranodec
numba==0.59.0
    # via pyod
numpy==1.26.4
    # via
    #   contourpy
    #   matplotlib
    #   numba
    #   outlier-utils
    #   pandas
    #   patsy
    #   pyarrow
    #   pyod
    #   pythresh
    #   scikit-learn
    #   scipy
    #   sranodec
    #   statsmodels
orjson==3.9.15
    # via -r requirements.in
outlier-utils==0.0.5
    # via -r requirements.in
packaging==24.0
    # via
    #   matplotlib
    #   statsmodels
paho-mqtt==1.6.1
    # via -r requirements.in
pandas==2.2.1
    # via
    #   -r requirements.in
    #   statsmodels
patsy==0.5.6
    # via statsmodels
pillow==10.2.0
    # via matplotlib
platformdirs==4.2.1
    # via zeep
prompt-toolkit==3.0.43
    # via click-repl
pyarrow==15.0.1
    # via -r requirements.in
pyod==1.1.3
    # via
    #   -r requirements.in
    #   pythresh
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pythresh==0.3.6
    # via -r requirements.in
pytz==2024.1
    # via
    #   pandas
    #   zeep
redis==5.0.3
    # via
    #   -r requirements.in
    #   celery
requests==2.31.0
    # via
    #   requests-file
    #   requests-toolbelt
    #   zeep
requests-file==2.0.0
    # via zeep
requests-toolbelt==1.0.0
    # via zeep
scikit-learn==1.4.1.post1
    # via
    #   pyod
    #   pythresh
scipy==1.12.0
    # via
    #   outlier-utils
    #   pyod
    #   pythresh
    #   scikit-learn
    #   sranodec
    #   statsmodels
six==1.16.0
    # via
    #   isodate
    #   patsy
    #   pyod
    #   python-dateutil
sranodec==0.0.1
    # via -r requirements.in
statsmodels==0.14.1
    # via -r requirements.in
threadpoolctl==3.3.0
    # via scikit-learn
typing-extensions==4.10.0
    # via kombu
tzdata==2024.1
    # via
    #   celery
    #   pandas
urllib3==2.2.1
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
zeep==4.2.1
    # via -r requirements.in
zipp==3.18.1
    # via importlib-resources
